backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pycryptodome>=3.19.0
numpy>=1.24.0
opencv-python-headless>=4.8.0
slowapi>=0.1.9
